seperate class for batching
