+++ tests/test_telemetry.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timedelta
 import time
-from utils import context, interfaces, missing_feature, bug, released, flaky, irrelevant, weblog, scenarios
+from utils import context, interfaces, missing_feature, bug, released, flaky, irrelevant, weblog
 
 from utils.tools import logger
 from utils.interfaces._misc_validators import HeadersPresenceValidator, HeadersMatchValidator
@@ -370,26 +370,3 @@ class Test_Telemetry:
         for dependency, seen in seen_loaded_dependencies.items():
             if not seen:
                 raise Exception(dependency + " not recieved in app-dependencies-loaded message")
-
-    def setup_message_batch_event_order(self):
-        weblog.get("/load_dependency")
-        weblog.get("/enable_integration")
-        weblog.get("/enable_product")
-
-    @missing_feature(
-        context.library in ("dotnet", "nodejs", "java", "python"),
-        reason="DD_FORCE_BATCHING_ENABLE is not implemented yet.",
-    )
-    @scenarios.telemetry_message_batch_event_order
-    def test_message_batch_event_order(self):
-        """Test that the events in message-batch are in chronological order"""
-        eventslist = []
-        for data in interfaces.library.get_telemetry_data():
-            content = data["request"]["content"]
-            eventslist.append(content.get("request_type"))
-
-        assert (
-            eventslist.index("app-dependencies-loaded")
-            < eventslist.index("app-integrations-change")
-            < eventslist.index("app-product-change")
-        ), "Events in message-batch are not in chronological order of event triggered"
